refactor: remove unused code and consolidate types
- Delete obsolete sidebar components and CSS modules
- Remove unused constants and imports
- Consolidate CreatePelanggaranInput type to api.ts
- Remove Settings menu without implementation

Cleanup removes ~370 lines of dead code
