--- src/components/layout/SidebarManagerItems.tsx
@@ -1,3 +1,0 @@
-import { Database } from 'lucide-react';
-
-export const managerItems = [{ label: 'Database', href: '/database', icon: Database }];
