+++ src/constants/index.ts
@@ -2,23 +2,4 @@
  * Application constants
  */
 
-export const DAYS = ['Senin', 'Selasa', 'Rabu', 'Kamis', 'Jumat', 'Sabtu', 'Minggu'] as const;
-export const DAYS_UPPER = ['SENIN', 'SELASA', 'RABU', 'KAMIS', 'JUMAT', 'SABTU'] as const;
-
-export const COLORS = {
-  primary: 'var(--primary)',
-  secondary: 'var(--secondary)',
-  success: 'var(--success)',
-  danger: 'var(--danger)',
-} as const;
-
-export const CHART_COLORS = [
-  '#8884d8',
-  '#82ca9d',
-  '#ffc658',
-  '#ff7300',
-  '#0088FE',
-  '#00C49F',
-] as const;
-
 export const ACTIVE_YEARS_THRESHOLD = 6;
